feat: minor changes to layout of group page
