+++ frontend/src/app/home/page.tsx
@@ -39,28 +39,8 @@ import Button from "@/components/buttons/Button";
 import Map from "@/components/Map";
 import { useGroupContext } from "../../../contexts/GroupContext";
 import { useRouter } from "next/navigation";
-import { signOut } from "next-auth/react";
-import { userApi } from "../../../lib/axios/userAPI";
+
 import HopinLogo from "../ui/hopin-logo";
-
-export const handleSignOut = async () => {
-  try {
-    // Clear both NextAuth and custom JWT tokens
-    await Promise.all([
-      userApi.signOut(), // Clear custom JWT tokens
-      signOut({ redirect: false }), // Clear NextAuth session without auto-redirect
-    ]);
-
-    console.log("User signed out successfully from all sessions");
-    // Navigate to home page
-    window.location.href = "/";
-  } catch (error) {
-    console.error("Error during unified sign out:", error);
-    // Even if there's an error, try to navigate away
-    alert("Sign out may have failed. Redirecting to home page.");
-    window.location.href = "/";
-  }
-};
 
 export default function HomePage() {
   const { createGroup } = useGroupContext();
@@ -73,25 +53,7 @@ export default function HomePage() {
 
   return (
     <>
-      <div className="flex flex-col md:flex-row justify-between p-2 items-center md:px-20 md:py-6  gap-2 md:gap-4">
-        <HopinLogo />
-
-        <div className="flex items-center gap-2 md:gap-4 justify-center">
-          <button
-            onClick={handleSignOut}
-            className=" text-xs md:text-sm font-bold border-1 p-2 rounded-sm"
-          >
-            Log Out
-          </button>
-
-          <button
-            onClick={() => router.push("/profile")}
-            className="text-xs md:text-sm font-bold border-1 p-2 rounded-sm"
-          >
-            Profile
-          </button>
-        </div>
-      </div>
+      <HopinLogo></HopinLogo>
       <div className="flex flex-col items-center gap-8 p-4 ">
         <div className="relative w-full h-[70vh] aspect-video">
           <Map />
